CI: Fix CI issue reporting
The native-tests job is run using a matrix which results in multiple
instances of the job being run per workflow run. This resulted in
multiple reports being generated as well. So for instance if the
native-tests job would run 6 times, with only half of them passing the
workflow could end up opening and closing the corresponding GH issue 6
times, while we would like it to only open it once if closed or post a
message that the workflow is still failing if already open. By moving
the report logic to a separate job that depends on the native-tests job
and reports the aggregated result of all the instances of the latter we
achieve our goal.