Retransform the verification map to use Providers as keys
This avoids reliance on precise instance (reference) equality
which we cannot support in the FIPS case as the SunPKCS11-NSS-FIPS
provider needs to be reinitialized at runtime.

Instead of using IdentityWrapper as key in the verification map
we transform it to a map using Provider instances as keys with
the same verification data. In the verification code we account
for this in JceSecurityUtil.

Fixes: MANDREL-98