Fix class-path order when Manifests include "Class-Path" attributes
`destination` is the classpath GraalVM creates from parsing the jar
files passed to it. The ordering of the entries in the classpath is
important. Jar files themselves should be added before the entries of
their corresponding "Class-Path" attributes.

At the same time we want to process the contents of
`META-INF/native-image` of the entries in the "Class-Path" before
processing the contents of `META-INF/native-image` of the jar itself.
This is required to ensure that the jar `META-INF/native-image` contents
can override those of the entries in its "Class-Path".

As a result I chose to optimistically add the path to the classpath (aka
`destination`) and remove it afterwards if deemed necessary.

Fixes https://github.com/oracle/graal/issues/7677

(cherry picked from commit 5c0bfbabb67117884ce1f4093347deea03a48354)