[CI] No longer build Mandrel with graal/master
Graal master switched to an incompatible (to OpenJDK) version as a base
JDK. It now needs Labs OpenJDK 25 with extra patches. There is not point
in trying to test the Mandrel config (which uses stock OpenJDK as base
JDK) with graal/master as it would fail.

We keep the configuration which uses LabsJDK.